[PAY-3198] New inbox settings modal layout
