+++ packages/web/src/components/inbox-settings-modal/InboxSettingsModalLegacy.module.css
